multimodal-chat/pyproject.toml
[project]
name = "multimodal-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.2.0",
    "ollama>=0.4.7",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
]
